+++ src/main/resources/constants/SettingsKeys.ts
@@ -9,5 +9,5 @@ export namespace SettingsKeys {
   export const UPDATE_CHECK = "updateCheck";
   export const UPDATE_CHECK_SHOW_NOTIFICATIONS = "updateCheck.showNotifications";
   export const AUTO_SAVE = "autoSave";
-  export const AUTO_SAVE_ENABLED = "autoSave.enabled";
+  export const AUTO_SAVE_TRIGGER = "autoSave.trigger";
 }